requirements.txt
Django
djangorestframework
gunicorn
drf-spectacular